created lib and exe
